requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
faiss-cpu==1.9.0
fastapi==0.115.4
fonttools==4.54.1
h11==0.14.0
huggingface_hub==0.26.2
idna==3.10
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.1
kiwisolver==1.4.7
matplotlib==3.9.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
networkx==3.4.2
numpy==1.26.2
packaging==23.2
parso==0.8.3
pillow==11.0.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
pure-eval==0.2.2
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.17.2
pyparsing==3.2.0
python-dateutil==2.8.2
pyzmq==25.1.2
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.2
tornado==6.4
traitlets==5.14.0
typing_extensions==4.12.2
uvicorn==0.32.0
wcwidth==0.2.12
sentence_transformers==3.2.1
einops==0.8.0